requirements.txt
Flask
bcrypt
Jinja2
PyJWT
pymongo
python-dotenv
openai
torch
tavily-python
transformers
accelerate
safetensors
invisible_watermark
peft
diffusers